fix: staking pools useEffect functions
