+++ src/views/Pools/components/PoolCard/AprRow.tsx
@@ -22,714 +22,20 @@ interface AprRowProps {
   isAutoVault?: boolean
   compoundFrequency?: number
   performanceFee?: number
+  rewardPerBlock?: string
 }
 
-const AprRow: React.FC<AprRowProps> = ({ pool, isAutoVault = false, compoundFrequency = 1, performanceFee = 0 }) => {
+const AprRow: React.FC<AprRowProps> = ({
+  pool,
+  isAutoVault = false,
+  compoundFrequency = 1,
+  performanceFee = 0,
+  rewardPerBlock,
+}) => {
   const { t } = useTranslation()
   const { stakingToken, earningToken, totalStaked, isFinished, tokenPerBlock } = pool
   const web3 = getWeb3NoAccount()
   // const newWeb3 = new Web3(Web3.givenProvider)
-
-  const [rewardPerBlock, setRewardPerBlock] = useState('')
-
-  const getRewardPerBlock = async () => {
-    if (pool) {
-      try {
-        if (
-          pool.poolCategory === '30DayLock' ||
-          pool.poolCategory === '60DayLock' ||
-          pool.poolCategory === '90DayLock'
-        ) {
-          const abi = [
-            { inputs: [], stateMutability: 'nonpayable', type: 'constructor' },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: false, internalType: 'address', name: 'tokenRecovered', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'AdminTokenRecovery',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'user', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'ClaimReward',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'user', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'Deposit',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'user', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'EmergencyWithdraw',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [{ indexed: false, internalType: 'uint256', name: 'poolLimitPerUser', type: 'uint256' }],
-              name: 'NewPoolLimit',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [{ indexed: false, internalType: 'uint256', name: 'rewardPerBlock', type: 'uint256' }],
-              name: 'NewRewardPerBlock',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: false, internalType: 'uint256', name: 'startBlock', type: 'uint256' },
-                { indexed: false, internalType: 'uint256', name: 'endBlock', type: 'uint256' },
-              ],
-              name: 'NewStartAndEndBlocks',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'previousOwner', type: 'address' },
-                { indexed: true, internalType: 'address', name: 'newOwner', type: 'address' },
-              ],
-              name: 'OwnershipTransferred',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'user', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'Withdraw',
-              type: 'event',
-            },
-            {
-              inputs: [],
-              name: 'PRECISION_FACTOR',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'SMART_CHEF_FACTORY',
-              outputs: [{ internalType: 'address', name: '', type: 'address' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'accTokenPerShare',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: '', type: 'address' }],
-              name: 'addressEndLockTime',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'bonusEndBlock',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            { inputs: [], name: 'claimReward', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-            {
-              inputs: [{ internalType: 'uint256', name: '_amount', type: 'uint256' }],
-              name: 'deposit',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'uint256', name: '_amount', type: 'uint256' }],
-              name: 'emergencyRewardWithdraw',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            { inputs: [], name: 'emergencyWithdraw', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-            {
-              inputs: [],
-              name: 'endLockTime',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: '_user', type: 'address' }],
-              name: 'getRemainingLockTime',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'hasAllRewardDistributedByAdmin',
-              outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'hasSavedPendingRewardUpdatedByAdmin',
-              outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'hasUserLimit',
-              outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [
-                { internalType: 'contract IBEP20', name: '_stakedToken', type: 'address' },
-                { internalType: 'contract IBEP20', name: '_rewardToken', type: 'address' },
-                { internalType: 'uint256', name: '_rewardPerBlock', type: 'uint256' },
-                { internalType: 'uint256', name: '_startBlock', type: 'uint256' },
-                { internalType: 'uint256', name: '_bonusEndBlock', type: 'uint256' },
-                { internalType: 'uint256', name: '_lockTime', type: 'uint256' },
-                { internalType: 'uint256', name: '_poolLimitPerUser', type: 'uint256' },
-                { internalType: 'address', name: '_admin', type: 'address' },
-              ],
-              name: 'initialize',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'isInitialized',
-              outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'lastRewardBlock',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'lockTime',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'numberOfClaimCurrentAndTotalPendingReward',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'numberOfClaimSavedPendingReward',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'owner',
-              outputs: [{ internalType: 'address', name: '', type: 'address' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: '_user', type: 'address' }],
-              name: 'pendingReward',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'poolLimitPerUser',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [
-                { internalType: 'address', name: '_tokenAddress', type: 'address' },
-                { internalType: 'uint256', name: '_tokenAmount', type: 'uint256' },
-              ],
-              name: 'recoverWrongTokens',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            { inputs: [], name: 'renounceOwnership', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-            {
-              inputs: [],
-              name: 'rewardPerBlock',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'rewardToken',
-              outputs: [{ internalType: 'contract IBEP20', name: '', type: 'address' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'stakedToken',
-              outputs: [{ internalType: 'contract IBEP20', name: '', type: 'address' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'startBlock',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            { inputs: [], name: 'stopReward', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-            {
-              inputs: [{ internalType: 'address', name: '', type: 'address' }],
-              name: 'temporaryPendingReward',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: 'newOwner', type: 'address' }],
-              name: 'transferOwnership',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [
-                { internalType: 'bool', name: '_hasUserLimit', type: 'bool' },
-                { internalType: 'uint256', name: '_poolLimitPerUser', type: 'uint256' },
-              ],
-              name: 'updatePoolLimitPerUser',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'uint256', name: '_rewardPerBlock', type: 'uint256' }],
-              name: 'updateRewardPerBlock',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [
-                { internalType: 'uint256', name: '_startBlock', type: 'uint256' },
-                { internalType: 'uint256', name: '_bonusEndBlock', type: 'uint256' },
-              ],
-              name: 'updateStartAndEndBlocks',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: '', type: 'address' }],
-              name: 'userInfo',
-              outputs: [
-                { internalType: 'uint256', name: 'amount', type: 'uint256' },
-                { internalType: 'uint256', name: 'rewardDebt', type: 'uint256' },
-              ],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'uint256', name: '_amount', type: 'uint256' }],
-              name: 'withdraw',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-          ]
-          const contract = new web3.eth.Contract(abi as unknown as AbiItem, getAddress(pool.contractAddress))
-          // console.log('reward contract', contract)
-
-          const rpb = await contract.methods.rewardPerBlock().call()
-          setRewardPerBlock(web3.utils.fromWei(rpb))
-        } else if (pool.poolCategory === 'Core') {
-          const abi = [
-            { inputs: [], stateMutability: 'nonpayable', type: 'constructor' },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: false, internalType: 'address', name: 'tokenRecovered', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'AdminTokenRecovery',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'user', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'ClaimReward',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'user', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'Deposit',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'user', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'EmergencyWithdraw',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [{ indexed: false, internalType: 'uint256', name: 'poolLimitPerUser', type: 'uint256' }],
-              name: 'NewPoolLimit',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [{ indexed: false, internalType: 'uint256', name: 'rewardPerBlock', type: 'uint256' }],
-              name: 'NewRewardPerBlock',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: false, internalType: 'uint256', name: 'startBlock', type: 'uint256' },
-                { indexed: false, internalType: 'uint256', name: 'endBlock', type: 'uint256' },
-              ],
-              name: 'NewStartAndEndBlocks',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'previousOwner', type: 'address' },
-                { indexed: true, internalType: 'address', name: 'newOwner', type: 'address' },
-              ],
-              name: 'OwnershipTransferred',
-              type: 'event',
-            },
-            {
-              anonymous: false,
-              inputs: [
-                { indexed: true, internalType: 'address', name: 'user', type: 'address' },
-                { indexed: false, internalType: 'uint256', name: 'amount', type: 'uint256' },
-              ],
-              name: 'Withdraw',
-              type: 'event',
-            },
-            {
-              inputs: [],
-              name: 'PRECISION_FACTOR',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'SMART_CHEF_FACTORY',
-              outputs: [{ internalType: 'address', name: '', type: 'address' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'accTokenPerShare',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: '', type: 'address' }],
-              name: 'addressEndLockTime',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'bonusEndBlock',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            { inputs: [], name: 'claimReward', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-            {
-              inputs: [{ internalType: 'uint256', name: '_amount', type: 'uint256' }],
-              name: 'deposit',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'uint256', name: '_amount', type: 'uint256' }],
-              name: 'emergencyRewardWithdraw',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            { inputs: [], name: 'emergencyWithdraw', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-            {
-              inputs: [],
-              name: 'endLockTime',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: '_user', type: 'address' }],
-              name: 'getRemainingLockTime',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'hasAllRewardDistributedByAdmin',
-              outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'hasSavedPendingRewardUpdatedByAdmin',
-              outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'hasUserLimit',
-              outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [
-                { internalType: 'contract IBEP20', name: '_stakedToken', type: 'address' },
-                { internalType: 'contract IBEP20', name: '_rewardToken', type: 'address' },
-                { internalType: 'uint256', name: '_rewardPerBlock', type: 'uint256' },
-                { internalType: 'uint256', name: '_startBlock', type: 'uint256' },
-                { internalType: 'uint256', name: '_bonusEndBlock', type: 'uint256' },
-                { internalType: 'uint256', name: '_lockTime', type: 'uint256' },
-                { internalType: 'uint256', name: '_poolLimitPerUser', type: 'uint256' },
-                { internalType: 'address', name: '_admin', type: 'address' },
-              ],
-              name: 'initialize',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'isInitialized',
-              outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'lastRewardBlock',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'lockTime',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'numberOfClaimCurrentAndTotalPendingReward',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'numberOfClaimSavedPendingReward',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'owner',
-              outputs: [{ internalType: 'address', name: '', type: 'address' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: '_user', type: 'address' }],
-              name: 'pendingReward',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'poolLimitPerUser',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [
-                { internalType: 'address', name: '_tokenAddress', type: 'address' },
-                { internalType: 'uint256', name: '_tokenAmount', type: 'uint256' },
-              ],
-              name: 'recoverWrongTokens',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            { inputs: [], name: 'renounceOwnership', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-            {
-              inputs: [],
-              name: 'rewardPerBlock',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'rewardToken',
-              outputs: [{ internalType: 'contract IBEP20', name: '', type: 'address' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'stakedToken',
-              outputs: [{ internalType: 'contract IBEP20', name: '', type: 'address' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [],
-              name: 'startBlock',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            { inputs: [], name: 'stopReward', outputs: [], stateMutability: 'nonpayable', type: 'function' },
-            {
-              inputs: [{ internalType: 'address', name: '', type: 'address' }],
-              name: 'temporaryPendingReward',
-              outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: 'newOwner', type: 'address' }],
-              name: 'transferOwnership',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [
-                { internalType: 'bool', name: '_hasUserLimit', type: 'bool' },
-                { internalType: 'uint256', name: '_poolLimitPerUser', type: 'uint256' },
-              ],
-              name: 'updatePoolLimitPerUser',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'uint256', name: '_rewardPerBlock', type: 'uint256' }],
-              name: 'updateRewardPerBlock',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [
-                { internalType: 'uint256', name: '_startBlock', type: 'uint256' },
-                { internalType: 'uint256', name: '_bonusEndBlock', type: 'uint256' },
-              ],
-              name: 'updateStartAndEndBlocks',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'address', name: '', type: 'address' }],
-              name: 'userInfo',
-              outputs: [
-                { internalType: 'uint256', name: 'amount', type: 'uint256' },
-                { internalType: 'uint256', name: 'rewardDebt', type: 'uint256' },
-              ],
-              stateMutability: 'view',
-              type: 'function',
-            },
-            {
-              inputs: [{ internalType: 'uint256', name: '_amount', type: 'uint256' }],
-              name: 'withdraw',
-              outputs: [],
-              stateMutability: 'nonpayable',
-              type: 'function',
-            },
-          ]
-          const contract = new web3.eth.Contract(abi as unknown as AbiItem, getAddress(pool.contractAddress))
-          // console.log('reward contract', contract)
-
-          const rpb = await contract.methods.rewardPerBlock().call()
-          setRewardPerBlock(web3.utils.fromWei(rpb))
-        }
-      } catch (error) {
-        // console.log('stake action', error)
-      }
-    }
-  }
-
-  useEffect(() => {
-    getRewardPerBlock()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [pool])
 
   const tooltipContent = isAutoVault
     ? t('APY includes compounding, APR doesn’t. This pool’s SOKU is compounded automatically, so we show APY.')
